Refactor Last.fm logic to dedicated LastFMViewModel
- Move all Last.fm authentication, user info, and scrobbling logic from BaseViewModel to a new LastFMViewModel for better modularity and MVVM separation.
- Update constructors for BaseViewModel, EditorViewModel, and SettingsViewModel to use DI for ILastfmService.
- Replace all DisplayAlert calls with DisplayAlertAsync for async consistency.
- Update animation extension methods to use new *Async MAUI APIs.
- Use SongsInAlbum property directly in album stats utilities for efficiency.
- Bump app version to 1.9.0.
- Add BoxView style and DevExpress namespace to Styles.xaml.
- Improve code clarity, fix typos, and enhance error handling and navigation in ViewModels.
- Strengthen adherence to MVVM and DI patterns for maintainability.
